+++ Vodovoz/ReportsParameters/CashFlow.cs
@@ -246,7 +246,7 @@ namespace Vodovoz.Reports
 				reportInfo.Parameters.Add("cash_subdivisions_name", cashSubdivisionsName);
 			}
 
-			var cashCategoryParametersProvider = new OrganisationCashTransferDocumentCashCategoryParametersProvider(SingletonParametersProvider.Instance);
+			var cashCategoryParametersProvider = new OrganizationCashTransferDocumentParametersProvider(SingletonParametersProvider.Instance);
 			reportInfo.Parameters.Add("cash_income_category_transfer_id", cashCategoryParametersProvider.CashIncomeCategoryTransferId);
 			reportInfo.Parameters.Add("cash_expense_category_transfer_id", cashCategoryParametersProvider.CashExpenseCategoryTransferId);
