+++ Vodovoz/gtk-gui/Vodovoz.Views.Cash.OrganizationCashTransferDocumentView.cs
@@ -1,0 +1,314 @@
+
+// This file has been generated by the GUI designer. Do not modify.
+namespace Vodovoz.Views.Cash
+{
+	public partial class OrganizationCashTransferDocumentView
+	{
+		private global::Gtk.VBox vboxDialog;
+
+		private global::Gtk.HBox hboxDialogButtons;
+
+		private global::Gamma.GtkWidgets.yButton buttonSave;
+
+		private global::Gamma.GtkWidgets.yButton buttonCancel;
+
+		private global::Gtk.Table tableState;
+
+		private global::Gtk.Label labelAuthor;
+
+		private global::Gamma.GtkWidgets.yLabel ylabelAuthor;
+
+		private global::Gtk.HSeparator hseparator2;
+
+		private global::Gtk.HBox hboxDate;
+
+		private global::Gtk.Label labelDocumentDate;
+
+		private global::Gamma.Widgets.yDatePicker ydatepickerDocumentDate;
+
+		private global::Gtk.HBox hboxTransferInfo;
+
+		private global::Gtk.Table tableFrom;
+
+		private global::Gtk.Label labelCashOrganizationFrom;
+
+		private global::QS.Widgets.GtkUI.SpecialListComboBox speciallistCmbOrganisationsFrom;
+
+		private global::Gtk.Table tableTo;
+
+		private global::Gtk.Label labelCashOrganizationTo;
+
+		private global::QS.Widgets.GtkUI.SpecialListComboBox speciallistCmbOrganisationsTo;
+
+		private global::Gtk.HSeparator hseparator1;
+
+		private global::Gtk.HBox hbox6;
+
+		private global::Gtk.Label labelSum;
+
+		private global::Gamma.GtkWidgets.ySpinButton yspinMoney;
+
+		private global::Gtk.Label labelComment;
+
+		private global::Gtk.ScrolledWindow GtkScrolledWindow2;
+
+		private global::Gamma.GtkWidgets.yTextView ytextviewComment;
+
+		protected virtual void Build()
+		{
+			global::Stetic.Gui.Initialize(this);
+			// Widget Vodovoz.Views.Cash.OrganizationCashTransferDocumentView
+			global::Stetic.BinContainer.Attach(this);
+			this.Name = "Vodovoz.Views.Cash.OrganizationCashTransferDocumentView";
+			// Container child Vodovoz.Views.Cash.OrganizationCashTransferDocumentView.Gtk.Container+ContainerChild
+			this.vboxDialog = new global::Gtk.VBox();
+			this.vboxDialog.Name = "vboxDialog";
+			this.vboxDialog.Spacing = 6;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.hboxDialogButtons = new global::Gtk.HBox();
+			this.hboxDialogButtons.Name = "hboxDialogButtons";
+			this.hboxDialogButtons.Spacing = 6;
+			// Container child hboxDialogButtons.Gtk.Box+BoxChild
+			this.buttonSave = new global::Gamma.GtkWidgets.yButton();
+			this.buttonSave.CanFocus = true;
+			this.buttonSave.Name = "buttonSave";
+			this.buttonSave.UseUnderline = true;
+			this.buttonSave.Label = global::Mono.Unix.Catalog.GetString("Сохранить");
+			global::Gtk.Image w1 = new global::Gtk.Image();
+			w1.Pixbuf = global::Stetic.IconLoader.LoadIcon(this, "gtk-save", global::Gtk.IconSize.Menu);
+			this.buttonSave.Image = w1;
+			this.hboxDialogButtons.Add(this.buttonSave);
+			global::Gtk.Box.BoxChild w2 = ((global::Gtk.Box.BoxChild)(this.hboxDialogButtons[this.buttonSave]));
+			w2.Position = 0;
+			w2.Expand = false;
+			w2.Fill = false;
+			// Container child hboxDialogButtons.Gtk.Box+BoxChild
+			this.buttonCancel = new global::Gamma.GtkWidgets.yButton();
+			this.buttonCancel.CanFocus = true;
+			this.buttonCancel.Name = "buttonCancel";
+			this.buttonCancel.UseUnderline = true;
+			this.buttonCancel.Label = global::Mono.Unix.Catalog.GetString("Отменить");
+			global::Gtk.Image w3 = new global::Gtk.Image();
+			w3.Pixbuf = global::Stetic.IconLoader.LoadIcon(this, "gtk-revert-to-saved", global::Gtk.IconSize.Menu);
+			this.buttonCancel.Image = w3;
+			this.hboxDialogButtons.Add(this.buttonCancel);
+			global::Gtk.Box.BoxChild w4 = ((global::Gtk.Box.BoxChild)(this.hboxDialogButtons[this.buttonCancel]));
+			w4.Position = 1;
+			w4.Expand = false;
+			w4.Fill = false;
+			this.vboxDialog.Add(this.hboxDialogButtons);
+			global::Gtk.Box.BoxChild w5 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.hboxDialogButtons]));
+			w5.Position = 0;
+			w5.Expand = false;
+			w5.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.tableState = new global::Gtk.Table(((uint)(1)), ((uint)(2)), false);
+			this.tableState.Name = "tableState";
+			this.tableState.RowSpacing = ((uint)(6));
+			this.tableState.ColumnSpacing = ((uint)(6));
+			// Container child tableState.Gtk.Table+TableChild
+			this.labelAuthor = new global::Gtk.Label();
+			this.labelAuthor.Name = "labelAuthor";
+			this.labelAuthor.Xalign = 1F;
+			this.labelAuthor.LabelProp = global::Mono.Unix.Catalog.GetString("Автор:");
+			this.tableState.Add(this.labelAuthor);
+			global::Gtk.Table.TableChild w6 = ((global::Gtk.Table.TableChild)(this.tableState[this.labelAuthor]));
+			w6.XOptions = ((global::Gtk.AttachOptions)(4));
+			w6.YOptions = ((global::Gtk.AttachOptions)(4));
+			// Container child tableState.Gtk.Table+TableChild
+			this.ylabelAuthor = new global::Gamma.GtkWidgets.yLabel();
+			this.ylabelAuthor.Name = "ylabelAuthor";
+			this.ylabelAuthor.Xalign = 0F;
+			this.tableState.Add(this.ylabelAuthor);
+			global::Gtk.Table.TableChild w7 = ((global::Gtk.Table.TableChild)(this.tableState[this.ylabelAuthor]));
+			w7.LeftAttach = ((uint)(1));
+			w7.RightAttach = ((uint)(2));
+			w7.XOptions = ((global::Gtk.AttachOptions)(4));
+			w7.YOptions = ((global::Gtk.AttachOptions)(4));
+			this.vboxDialog.Add(this.tableState);
+			global::Gtk.Box.BoxChild w8 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.tableState]));
+			w8.Position = 1;
+			w8.Expand = false;
+			w8.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.hseparator2 = new global::Gtk.HSeparator();
+			this.hseparator2.Name = "hseparator2";
+			this.vboxDialog.Add(this.hseparator2);
+			global::Gtk.Box.BoxChild w9 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.hseparator2]));
+			w9.Position = 2;
+			w9.Expand = false;
+			w9.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.hboxDate = new global::Gtk.HBox();
+			this.hboxDate.Name = "hboxDate";
+			this.hboxDate.Spacing = 6;
+			// Container child hboxDate.Gtk.Box+BoxChild
+			this.labelDocumentDate = new global::Gtk.Label();
+			this.labelDocumentDate.Name = "labelDocumentDate";
+			this.labelDocumentDate.LabelProp = global::Mono.Unix.Catalog.GetString("Дата:");
+			this.hboxDate.Add(this.labelDocumentDate);
+			global::Gtk.Box.BoxChild w10 = ((global::Gtk.Box.BoxChild)(this.hboxDate[this.labelDocumentDate]));
+			w10.Position = 0;
+			w10.Expand = false;
+			w10.Fill = false;
+			// Container child hboxDate.Gtk.Box+BoxChild
+			this.ydatepickerDocumentDate = new global::Gamma.Widgets.yDatePicker();
+			this.ydatepickerDocumentDate.WidthRequest = 300;
+			this.ydatepickerDocumentDate.Events = ((global::Gdk.EventMask)(256));
+			this.ydatepickerDocumentDate.Name = "ydatepickerDocumentDate";
+			this.ydatepickerDocumentDate.WithTime = false;
+			this.ydatepickerDocumentDate.Date = new global::System.DateTime(0);
+			this.ydatepickerDocumentDate.IsEditable = true;
+			this.ydatepickerDocumentDate.AutoSeparation = false;
+			this.hboxDate.Add(this.ydatepickerDocumentDate);
+			global::Gtk.Box.BoxChild w11 = ((global::Gtk.Box.BoxChild)(this.hboxDate[this.ydatepickerDocumentDate]));
+			w11.Position = 1;
+			w11.Expand = false;
+			this.vboxDialog.Add(this.hboxDate);
+			global::Gtk.Box.BoxChild w12 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.hboxDate]));
+			w12.Position = 3;
+			w12.Expand = false;
+			w12.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.hboxTransferInfo = new global::Gtk.HBox();
+			this.hboxTransferInfo.Name = "hboxTransferInfo";
+			this.hboxTransferInfo.Spacing = 6;
+			// Container child hboxTransferInfo.Gtk.Box+BoxChild
+			this.tableFrom = new global::Gtk.Table(((uint)(1)), ((uint)(3)), false);
+			this.tableFrom.Name = "tableFrom";
+			this.tableFrom.RowSpacing = ((uint)(6));
+			this.tableFrom.ColumnSpacing = ((uint)(6));
+			// Container child tableFrom.Gtk.Table+TableChild
+			this.labelCashOrganizationFrom = new global::Gtk.Label();
+			this.labelCashOrganizationFrom.Name = "labelCashOrganizationFrom";
+			this.labelCashOrganizationFrom.Xalign = 1F;
+			this.labelCashOrganizationFrom.LabelProp = global::Mono.Unix.Catalog.GetString("Организация откуда:");
+			this.tableFrom.Add(this.labelCashOrganizationFrom);
+			global::Gtk.Table.TableChild w13 = ((global::Gtk.Table.TableChild)(this.tableFrom[this.labelCashOrganizationFrom]));
+			w13.XOptions = ((global::Gtk.AttachOptions)(4));
+			w13.YOptions = ((global::Gtk.AttachOptions)(4));
+			// Container child tableFrom.Gtk.Table+TableChild
+			this.speciallistCmbOrganisationsFrom = new global::QS.Widgets.GtkUI.SpecialListComboBox();
+			this.speciallistCmbOrganisationsFrom.Name = "speciallistCmbOrganisationsFrom";
+			this.speciallistCmbOrganisationsFrom.AddIfNotExist = false;
+			this.speciallistCmbOrganisationsFrom.DefaultFirst = false;
+			this.speciallistCmbOrganisationsFrom.ShowSpecialStateAll = true;
+			this.speciallistCmbOrganisationsFrom.ShowSpecialStateNot = false;
+			this.tableFrom.Add(this.speciallistCmbOrganisationsFrom);
+			global::Gtk.Table.TableChild w14 = ((global::Gtk.Table.TableChild)(this.tableFrom[this.speciallistCmbOrganisationsFrom]));
+			w14.LeftAttach = ((uint)(1));
+			w14.RightAttach = ((uint)(2));
+			w14.XOptions = ((global::Gtk.AttachOptions)(4));
+			w14.YOptions = ((global::Gtk.AttachOptions)(4));
+			this.hboxTransferInfo.Add(this.tableFrom);
+			global::Gtk.Box.BoxChild w15 = ((global::Gtk.Box.BoxChild)(this.hboxTransferInfo[this.tableFrom]));
+			w15.Position = 0;
+			// Container child hboxTransferInfo.Gtk.Box+BoxChild
+			this.tableTo = new global::Gtk.Table(((uint)(1)), ((uint)(3)), false);
+			this.tableTo.Name = "tableTo";
+			this.tableTo.RowSpacing = ((uint)(6));
+			this.tableTo.ColumnSpacing = ((uint)(6));
+			// Container child tableTo.Gtk.Table+TableChild
+			this.labelCashOrganizationTo = new global::Gtk.Label();
+			this.labelCashOrganizationTo.Name = "labelCashOrganizationTo";
+			this.labelCashOrganizationTo.Xalign = 1F;
+			this.labelCashOrganizationTo.LabelProp = global::Mono.Unix.Catalog.GetString("Организация куда:");
+			this.tableTo.Add(this.labelCashOrganizationTo);
+			global::Gtk.Table.TableChild w16 = ((global::Gtk.Table.TableChild)(this.tableTo[this.labelCashOrganizationTo]));
+			w16.XOptions = ((global::Gtk.AttachOptions)(4));
+			w16.YOptions = ((global::Gtk.AttachOptions)(4));
+			// Container child tableTo.Gtk.Table+TableChild
+			this.speciallistCmbOrganisationsTo = new global::QS.Widgets.GtkUI.SpecialListComboBox();
+			this.speciallistCmbOrganisationsTo.Name = "speciallistCmbOrganisationsTo";
+			this.speciallistCmbOrganisationsTo.AddIfNotExist = false;
+			this.speciallistCmbOrganisationsTo.DefaultFirst = false;
+			this.speciallistCmbOrganisationsTo.ShowSpecialStateAll = true;
+			this.speciallistCmbOrganisationsTo.ShowSpecialStateNot = false;
+			this.tableTo.Add(this.speciallistCmbOrganisationsTo);
+			global::Gtk.Table.TableChild w17 = ((global::Gtk.Table.TableChild)(this.tableTo[this.speciallistCmbOrganisationsTo]));
+			w17.LeftAttach = ((uint)(1));
+			w17.RightAttach = ((uint)(2));
+			w17.XOptions = ((global::Gtk.AttachOptions)(4));
+			w17.YOptions = ((global::Gtk.AttachOptions)(4));
+			this.hboxTransferInfo.Add(this.tableTo);
+			global::Gtk.Box.BoxChild w18 = ((global::Gtk.Box.BoxChild)(this.hboxTransferInfo[this.tableTo]));
+			w18.Position = 1;
+			this.vboxDialog.Add(this.hboxTransferInfo);
+			global::Gtk.Box.BoxChild w19 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.hboxTransferInfo]));
+			w19.Position = 4;
+			w19.Expand = false;
+			w19.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.hseparator1 = new global::Gtk.HSeparator();
+			this.hseparator1.Name = "hseparator1";
+			this.vboxDialog.Add(this.hseparator1);
+			global::Gtk.Box.BoxChild w20 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.hseparator1]));
+			w20.Position = 5;
+			w20.Expand = false;
+			w20.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.hbox6 = new global::Gtk.HBox();
+			this.hbox6.Name = "hbox6";
+			this.hbox6.Spacing = 6;
+			// Container child hbox6.Gtk.Box+BoxChild
+			this.labelSum = new global::Gtk.Label();
+			this.labelSum.Name = "labelSum";
+			this.labelSum.Xalign = 1F;
+			this.labelSum.LabelProp = global::Mono.Unix.Catalog.GetString("Сумма к отправке:");
+			this.hbox6.Add(this.labelSum);
+			global::Gtk.Box.BoxChild w21 = ((global::Gtk.Box.BoxChild)(this.hbox6[this.labelSum]));
+			w21.Position = 0;
+			w21.Expand = false;
+			w21.Fill = false;
+			// Container child hbox6.Gtk.Box+BoxChild
+			this.yspinMoney = new global::Gamma.GtkWidgets.ySpinButton(0D, 999999999D, 100D);
+			this.yspinMoney.CanDefault = true;
+			this.yspinMoney.CanFocus = true;
+			this.yspinMoney.Name = "yspinMoney";
+			this.yspinMoney.Adjustment.PageIncrement = 1000D;
+			this.yspinMoney.ClimbRate = 1D;
+			this.yspinMoney.Digits = ((uint)(2));
+			this.yspinMoney.Numeric = true;
+			this.yspinMoney.ValueAsDecimal = 0m;
+			this.yspinMoney.ValueAsInt = 0;
+			this.hbox6.Add(this.yspinMoney);
+			global::Gtk.Box.BoxChild w22 = ((global::Gtk.Box.BoxChild)(this.hbox6[this.yspinMoney]));
+			w22.Position = 1;
+			w22.Expand = false;
+			w22.Fill = false;
+			this.vboxDialog.Add(this.hbox6);
+			global::Gtk.Box.BoxChild w23 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.hbox6]));
+			w23.Position = 6;
+			w23.Expand = false;
+			w23.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.labelComment = new global::Gtk.Label();
+			this.labelComment.Name = "labelComment";
+			this.labelComment.Xalign = 0F;
+			this.labelComment.LabelProp = global::Mono.Unix.Catalog.GetString("Комментарий:");
+			this.vboxDialog.Add(this.labelComment);
+			global::Gtk.Box.BoxChild w24 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.labelComment]));
+			w24.Position = 7;
+			w24.Expand = false;
+			w24.Fill = false;
+			// Container child vboxDialog.Gtk.Box+BoxChild
+			this.GtkScrolledWindow2 = new global::Gtk.ScrolledWindow();
+			this.GtkScrolledWindow2.Name = "GtkScrolledWindow2";
+			this.GtkScrolledWindow2.ShadowType = ((global::Gtk.ShadowType)(1));
+			// Container child GtkScrolledWindow2.Gtk.Container+ContainerChild
+			this.ytextviewComment = new global::Gamma.GtkWidgets.yTextView();
+			this.ytextviewComment.CanFocus = true;
+			this.ytextviewComment.Name = "ytextviewComment";
+			this.GtkScrolledWindow2.Add(this.ytextviewComment);
+			this.vboxDialog.Add(this.GtkScrolledWindow2);
+			global::Gtk.Box.BoxChild w26 = ((global::Gtk.Box.BoxChild)(this.vboxDialog[this.GtkScrolledWindow2]));
+			w26.Position = 8;
+			this.Add(this.vboxDialog);
+			if ((this.Child != null))
+			{
+				this.Child.ShowAll();
+			}
+			this.Hide();
+		}
+	}
+}
