+++ Vodovoz/toolbars/cash.xml
@@ -8,6 +8,6 @@
         <toolitem name='ActionSelfdeliveryOrders' action='ActionSelfdeliveryOrders'/>
         <toolitem name='ActionCashTransferDocuments' action='ActionCashTransferDocuments'/>
         <toolitem name='ActionFuelTransferDocuments' action='ActionFuelTransferDocuments'/>
-        <toolitem name='ActionOrganisationCashTransferDocuments' action='ActionOrganisationCashTransferDocuments'/>
+        <toolitem name='ActionOrganizationCashTransferDocuments' action='ActionOrganizationCashTransferDocuments'/>
 	</toolbar>
 </ui>
