+++ VodovozBusiness/ConfigureDeletion.cs
@@ -1042,7 +1042,7 @@ namespace Vodovoz
 			DeleteConfig.AddHibernateDeleteInfo<CashRequestSumItem>()
 				.AddDeleteDependence<Expense>(x => x.CashRequestSumItem);
 
-			DeleteConfig.AddHibernateDeleteInfo<OrganisationCashTransferDocument>()
+			DeleteConfig.AddHibernateDeleteInfo<OrganizationCashTransferDocument>()
 				.AddDeleteCascadeDependence(x => x.OrganisationCashMovementOperationFrom)
 				.AddDeleteCascadeDependence(x => x.OrganisationCashMovementOperationTo);
