--- VodovozBusiness/Services/IOrganisationCashTransferDocumentCashCategoryParametersProvider.cs
@@ -1,8 +1,0 @@
-namespace Vodovoz.Services
-{
-    public interface IOrganisationCashTransferDocumentCashCategoryParametersProvider
-    {
-        int CashIncomeCategoryTransferId { get; }
-        int CashExpenseCategoryTransferId { get; }
-    }
-}
