+++ VodovozBusiness/Services/IOrganizationCashTransferDocumentParametersProvider.cs
@@ -1,0 +1,8 @@
+namespace Vodovoz.Services
+{
+    public interface IOrganizationCashTransferDocumentParametersProvider
+    {
+        int CashIncomeCategoryTransferId { get; }
+        int CashExpenseCategoryTransferId { get; }
+    }
+}
