--- VodovozViewModels/ViewModels/Cash/OrganisationCashTransferDocumentViewModel.cs
@@ -1,68 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using QS.DomainModel.Entity.EntityPermissions.EntityExtendedPermission;
-using QS.DomainModel.UoW;
-using QS.Navigation;
-using QS.Project.Domain;
-using QS.Services;
-using QS.ViewModels;
-using Vodovoz.Domain.Cash;
-using Vodovoz.Domain.Employees;
-using Vodovoz.Domain.Operations;
-using Vodovoz.Domain.Organizations;
-using Vodovoz.PermissionExtensions;
-using Vodovoz.Repositories.HumanResources;
-
-namespace Vodovoz.ViewModels.ViewModels.Cash
-{
-    public class OrganisationCashTransferDocumentViewModel : EntityTabViewModelBase<OrganisationCashTransferDocument>
-    {
-        public IEnumerable<Organization> Organizations { get; }
-        public bool CanEdit { get; }
-        public bool CanEditRectroactively { get; }
-        private readonly Employee author;
-
-        public OrganisationCashTransferDocumentViewModel(IEntityUoWBuilder uowBuilder, IUnitOfWorkFactory unitOfWorkFactory, ICommonServices commonServices, IEntityExtendedPermissionValidator entityExtendedPermissionValidator)
-            : base(uowBuilder, unitOfWorkFactory, commonServices)
-        {
-            Organizations = UoW.GetAll<Organization>();
-            author = EmployeeRepository.GetEmployeeForCurrentUser(UoW);
-
-            CanEditRectroactively = entityExtendedPermissionValidator.Validate(typeof(OrganisationCashTransferDocument), CommonServices.UserService.CurrentUserId, nameof(RetroactivelyClosePermission));
-            CanEdit = (Entity.Id == 0 && PermissionResult.CanCreate) ||
-                      (PermissionResult.CanUpdate && Entity.DocumentDate.Date == DateTime.Now.Date) ||
-                      CanEditRectroactively;
-
-            if (Entity.Id == 0)
-            {
-                Entity.DocumentDate = DateTime.Now;
-                Entity.Author = author;
-            }
-        }
-
-        public override bool Save(bool close)
-        {
-            if (!CanEdit || !Validate())
-                return false;
-
-            if (HasChanges)
-            {
-                var operationFrom = Entity.OrganisationCashMovementOperationFrom = Entity.OrganisationCashMovementOperationFrom ?? new OrganisationCashMovementOperation { OperationTime = DateTime.Now };
-                var operationTo = Entity.OrganisationCashMovementOperationTo = Entity.OrganisationCashMovementOperationTo ?? new OrganisationCashMovementOperation { OperationTime = DateTime.Now };
-
-                operationFrom.Organisation = Entity.OrganizationFrom;
-                operationFrom.Amount = -Entity.TransferedSum;
-
-                operationTo.Organisation = Entity.OrganizationTo;
-                operationTo.Amount = Entity.TransferedSum;
-
-                UoW.Save(operationFrom);
-                UoW.Save(operationTo);
-                UoW.Save();
-            }
-
-            Close(false, CloseSource.Save);
-            return true;
-        }
-    }
-}
